Fix validity window of fake cert in test_cert_reload
The test_cert_reload unit test uses a fake X.509 server certificate.
For the test to pass, the current date as reported by the system must
be betwen the certificate's "not valid before" and "not valid after"
dates. This was true when the test was first written, but the test
started failing when the original "not valid after" date of
2020-01-01 had passed.

Work around this by setting the "not valid before" date to the
distant past (2008-01-01) and the "not valid after" date to the
distant future (3020-01-01).

Our generation could not solve this. Maybe our
great-great-grandchildren will be wiser.

Fixes #2.